# Optimize LabelValues in metric proto