@SergeyKanzhelev 2019-11-01
> We declared that no, the value is explicitly unspecified if the LabelSet/Handle did not declare it.

From consumption standpoint - what's the difference of non-specified and empty value? Would one want to create a stacked chart where unspecified label and empty label are two different values? If empty is that important for this scenario - perhaps one can report it with the actual word "EMPTY"?

If there are some notes I can read thru to understand it - please let me know