@tigrannajaryan 2019-11-06
@SergeyKanzhelev @jmacd can we move forward with this change? It does not introduce new concepts, but merely follows what spec requires today.

I created a separate issue to clarify the need for "unspecified" lablels here: https://github.com/open-telemetry/opentelemetry-specification/issues/345

If the conclusion eventually is that we don't need the unspecified labels I will happily modify this proto to remove them in a future PR. For now I believe this PR is good as is.