@jmacd 2019-11-06
In hindsight, I'm not sure we should have written that into the specification. It was related to the concept of _required_ aggregation keys, which was softened in the final spec to _recommended_ aggregation keys. If a system is expecting precisely a fixed number of dimensions and those dimensions are not included in the export format, they are effectively unspecified. 

The reason we discussed explicitly unspecified values was to declare that the implementation would not be required to scan through the context for distributed correlation labels to satisfy the missing values. 

@bogdandrutu I'd be happy to strike this from the specification.