@tigrannajaryan 2019-11-06
> I'm fine with this change, but declaring empty strings and unspecified values to be the same would be a better optimization and easier to understand

@SergeyKanzhelev I agree and if spec decides to do that I will make the corresponding change here.