@tigrannajaryan 2019-11-06
> @SergeyKanzhelev I didn't say empty string--that I would reject. I said leaving out a key:value entirely (i.e., no key, no value) is effectively saying unspecified.

And that's exactly possible to do now, by omitting the key from `MetricDescriptor.label_keys`. The only downside to that is that when exporting metrics you will have to create distinct instances of `MetricDescriptor` for each combination of specified labels (as opposed to have a single `MetricDescriptor` for all instances of one metric that have different label values). It is doable and my gut feeling it is not a big deal since it is going to be rare.