@tigrannajaryan 2019-11-07
> Hard to follow this discussion. Can we summarize what is the current status? I personally prefer to first decide if we need to support unspecified or we can do what OpenMetrics does and treat unspecified as empty.

@bogdandrutu to summarize: I believe we can move forward and merge this PR since it is a straight improvement over current state and does not change any semantics, it is simply a performance optimization without functional changes.

We can separately discuss if we want to eliminate "unspecified label" concept as part of https://github.com/open-telemetry/opentelemetry-specification/issues/345 that I created for that discussion.

Depending on the decision resulting from that discussion I will make any necessary changes in this proto repository.