@tigrannajaryan 2019-11-08
@SergeyKanzhelev @jmacd I removed the ability to have unspecified value from this PR. Please have another look.