@SergeyKanzhelev 2019-11-09
@jmacd I'm merging this just to move forward. If you feel that we need to revisit this decision - please file an issue.